Adding a failing test for some odd behavior we came across.
If I want anyone to be able to :read objects of a Class but only :manage
a more specific subset I might do the following:

can :read  , MyClass
can :manage, MyClass, :owner_id => user.id

This works fine for authorization but when cancan tries to fetch things
for us in an :index it uses the {:owner_id => user.id} filter and I only see
objects that I can :manage.

This behavior seems counter-intuitive. I would expect the results
fetched for :index to include all the Objects that I can :read.