Remove comparison of class names
This was added by a [patch a couple of years
ago](https://github.com/ryanb/cancan/commit/79180de3722673388c1eb65fc09db5a72fe0b1a8),
apparently to fix a problem with class reloading. I'm not sure if that problem
still obtains but the string comparison is a performance bottleneck, and
removing it should not change the semantics of the method.