Change how to call can passing multiple subjects
During test to ensure that it was still possible to pass an object represented by a Hash, I saw that the way how can? receive multiple subjects needed change.

The way I could implement without loosing functionality, was to receive the array of subjects inside a hash labeled by the symbol :any

So the way how to pass multiple arrays now is
 can? :read, {:any => [object1, object2]}